Fixed critical bug in Umfrageergebnis Service
The order of the Vorschläge in the Umfrageergebnis UI did not correspond
to the order of the answers, which led to a faulty mapping being
displayed in the browser.

The Umfrageergebnis UI receives its Vorschläge and answers inside of an
Umfrageergebnis Form, which in turn is being constructed inside of the
Umfrageabstimmung Controller. Finally, the construction takes place
inside of the Umfrageergebnis Service.

Upon inspection, the Umfrageergebnis Service employs a Hash Map to store
the answers and the Vorschläge which they belong to. Since a Hash Map is
using the objects' hash for storage, the resulting order is very likely
to change as new elements are inserted.

The solution to this was to replace the Hash Map with a Linked Hash Map,
which retains the order in which elements are inserted.