Removed unnecessary for loop and conditional in UmfrageUebersichtService
The changed method checks if the Frist of an Umfrage is in the future or
in the past. However, the method is only called right after the Umfragen
are filtered by their Frist being in the past. So the conditional is
always only going in one branch, making it completely unnecessary